pyproject.toml
[project]
name = "mlops-pratice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "black==25.1.0",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "lightning[extra]==2.5.0.post0",
    "loguru==0.7.3",
    "pytest>=8.3.4",
    "rich==13.9.4",
    "tensorboard==2.18.0",
    "timm==1.0.14",
    "torch==2.6.0+cpu",
    "torchvision==0.21.0+cpu",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v -s --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
